refactor: move outbound helpers to shared utils
