+++ apps/api/tests/contract/test_s3_event_contract.py
@@ -198,7 +198,7 @@ async def test_should_reject_an_sns_subscription_confirmation_url_that_resolves_
     async with api_client_factory() as api_client:
         monkeypatch.setattr(socket, "getaddrinfo", resolve_private_address)
         pinned_http_module = importlib.import_module(
-            "shared.services.webhook.pinned_outbound_http"
+            "shared.utils.pinned_outbound_http"
         )
         monkeypatch.setattr(
             pinned_http_module.aiohttp,
