+++ packages/shared-python/shared/utils/pinned_outbound_http.py
